+++ Packages/com.girlydrink.credit-tool/Runtime/CreditComponent.cs
@@ -6,6 +6,7 @@ namespace GirlyDrink.CreditTool
     public class CreditComponent : MonoBehaviour, IEditorOnly
     {
         public string authorName = "";
+        public string authorStoreLink = "";
         public string assetName = "";
         public string assetLink = "";
     }